Basic functionality completed. What needs to be done: export to excel add a tutorial/help page statistics of recording behavior on connect tab

+++ View/Gui.cs
@@ -12,41 +12,70 @@ using WatchCommunication;
 
 public partial class Gui: Form
 {
-    private List<Student> students;
     private IRepository repository;
     private IAccessPoint accessPoint;
     private WatchPool watchPool;
+    private List<ListBox> studentListBoxes;
+    private int count;
 
     public Gui(IRepository repository, IAccessPoint accessPoint) {
         this.repository = repository;
         this.accessPoint = accessPoint;
+
         watchPool = new WatchPool();
+        count = 0;
+        studentListBoxes = new List<ListBox>();
         InitializeComponent();
     }
 
     private void Gui_Load(object sender, EventArgs e) {
-        students = repository.All().ToList();
-        deleteStudentsListBox.DataSource = students;
+        studentListBoxes.Add(deleteStudentsListBox);
+        studentListBoxes.Add(addedStudentsListBox);
+        RefreshStudentListBoxes();
+
+        RefreshAvailableStudentsListBox();
+        RefreshAvailableWatchesListBox();
+
         accessPoint.OnPacketRecieved += new Action<uint>(accessPoint_OnPacketRecieved);
-        accessPoint.OnLostConnection += () => {
-            accessPoint.Close();
-            new View.CouldNotConnectError(accessPoint).ShowDialog();
-        };
+        accessPoint.OnLostConnection += new Action(accessPoint_OnLostConnection);
+
+        connect();
+    }
+
+
+
+    private void connect() {
         try {
             accessPoint.Open();
         }
         catch {
             new View.CouldNotConnectError(accessPoint).ShowDialog();
         }
+        refreshConnectionDependentComponents();
     }
+
+    private void accessPoint_OnLostConnection() {
+        accessPoint.Close();
+        this.Invoke(new Action(() => this.Enabled = false));
+        var dialog = new View.CouldNotConnectError(accessPoint);
+        dialog.FormClosed += (x, y) => {
+            this.Invoke(new Action(() => this.Enabled = true));
+        };
+        dialog.ShowDialog();
+        refreshConnectionDependentComponents();
+    }
+
+
 
     private void accessPoint_OnPacketRecieved(uint packet) {
         if (watchPool.HasWatchWithIdentifier(packet)) {
             var watch = watchPool.WithIdentifier(packet);
             try {
                 var behavior = new Behavior();
-                watch.Student.Behaviors.Add(behavior);
-                repository.Commit();
+                if (watch.Student != null) {
+                    watch.Student.Behaviors.Add(behavior);
+                    repository.Commit();
+                }
             }
             catch {
                 repository.Rollback();
@@ -54,11 +83,78 @@ public partial class Gui: Form
             }
         }
         else {
+            watchPool.Add(new Watch { PacketIdentifier = packet });
+            RefreshAvailableWatchesListBox();
         }
-        updateLabel.Invoke(new Action(() => updateLabel.Text = DateTime.Now.ToString()));
+        this.Invoke(new Action(() => { count++; updateLabel.Text = count.ToString(); }));
+    }
+
+    #region refresh methods
+
+    private void RefreshStudentListBoxes() {
+        var students = repository.All().ToList();
+        foreach (var l in studentListBoxes) {
+            l.DataSource = students;
+            l.Refresh();
+        }
+    }
+
+    private void refreshConnectionDependentComponents() {
+        Action connected = () => {
+            isConnectedLabel.Visible = true;
+            isNotConnectedLabel.Visible = false;
+            connectButton.Visible = false;
+        };
+        Action notConnected = () => {
+            isConnectedLabel.Visible = false;
+            isNotConnectedLabel.Visible = true;
+            connectButton.Visible = true;
+        };
+
+        if (accessPoint.IsConnected())
+            this.Invoke(connected);
+        else
+            this.Invoke(notConnected);
+    }
+
+    private void RefreshAvailableWatchesListBox() {
+        this.Invoke(new Action(() => {
+            availableWatchesListBox.DataSource = watchPool.Where(w => w.Student == null).ToList();
+            availableWatchesListBox.Refresh();
+        }));
+    }
+
+    private void RefreshAvailableStudentsListBox() {
+        this.Invoke(new Action(() => {
+            var allStudents = repository.All().ToList();
+            var studentsWithWatches = watchPool.Select(w => w.Student).Where(s => s != null);
+            availableStudentsListBox.DataSource = allStudents.Except(studentsWithWatches).ToList();
+            availableStudentsListBox.Refresh();
+        }));
+    }
+
+    private void RefreshPairedStudentsListBox() {
+        this.Invoke(new Action(() => {
+            var studentsWithWatches = watchPool.Select(w => w.Student).Where(s => s != null).ToList();
+            pairedStudentsListBox.DataSource = studentsWithWatches;
+            pairedStudentsListBox.DataSource = studentsWithWatches;
+            availableStudentsListBox.Refresh();
+        }));
+    }
+
+    #endregion refresh methods
+
+    private static void UnknownFailureMessageBox() {
+        MessageBox.Show(
+            "An unknown error occured. Try closing Metamorphosis and reopening.",
+            "Failure",
+            MessageBoxButtons.OK,
+            MessageBoxIcon.Error
+        );
     }
 
     #region event handlers
+
     private void addStudentButton_Click(object sender, EventArgs e) {
         try {
             var student = new Student(
@@ -70,7 +166,6 @@ public partial class Gui: Form
             );
             repository.Add(student);
             repository.Commit();
-            students = repository.All().ToList();
             MessageBox.Show("Student succesfully added.",
                 "Success",
                 MessageBoxButtons.OK,
@@ -89,34 +184,75 @@ public partial class Gui: Form
             repository.Rollback();
             UnknownFailureMessageBox();
         }
+        firstNameTextBox.Clear();
+        lastNameTextBox.Clear();
+        teacherTextBox.Clear();
+        behaviorTextBox.Clear();
+        RefreshStudentListBoxes();
+        RefreshAvailableStudentsListBox();
     }
 
 
     private void deleteButton_Click(object sender, EventArgs e) {
-        IList<Student> selected = deleteStudentsListBox.SelectedItems.Cast<Student>().ToList();
-        try {
-            foreach (var s in selected)
-                repository.Delete(s);
-            repository.Commit();
-            MessageBox.Show("Students succesfully deleted.",
-            "Success",
-            MessageBoxButtons.OK,
-            MessageBoxIcon.Information);
+        var result = MessageBox.Show(
+                "Are you sure you want to delete the selected students?",
+                "Warning",
+                MessageBoxButtons.YesNo,
+                MessageBoxIcon.Warning);
+        if (result == DialogResult.Yes) {
+            IList<Student> selected = deleteStudentsListBox.SelectedItems.Cast<Student>().ToList();
+            try {
+                foreach (var s in selected)
+                    repository.Delete(s);
+                repository.Commit();
+                MessageBox.Show(
+                "Students succesfully deleted.",
+                "Success",
+                MessageBoxButtons.OK,
+                MessageBoxIcon.Information);
+            }
+            catch {
+                repository.Rollback();
+                UnknownFailureMessageBox();
+            }
         }
-        catch {
-            repository.Rollback();
-            UnknownFailureMessageBox();
+        RefreshStudentListBoxes();
+        RefreshAvailableStudentsListBox();
+    }
+
+    private void connectButton_Click(object sender, EventArgs e) {
+        connect();
+    }
+
+    #endregion
+
+    private void pairButton_Click(object sender, EventArgs e) {
+        var selectedStudent = (Student)availableStudentsListBox.SelectedItem;
+        var selectedWatch = (Watch)availableWatchesListBox.SelectedItem;
+        if (selectedWatch != null && selectedStudent != null) {
+            selectedWatch.Student = selectedStudent;
+            RefreshAvailableWatchesListBox();
+            RefreshAvailableStudentsListBox();
+            RefreshPairedStudentsListBox();
         }
 
     }
-    #endregion
 
-    private static void UnknownFailureMessageBox() {
-        MessageBox.Show(
-            "An unknown error occured. Try closing Metamorphosis and reopening.",
-            "Failure",
-            MessageBoxButtons.OK,
-            MessageBoxIcon.Error
-        );
+    private void unpairButton_Click(object sender, EventArgs e) {
+        var selectedStudent = (Student)pairedStudentsListBox.SelectedItem;
+        if (selectedStudent != null) {
+            var pairedWatch = watchPool
+                .Where(w => w.Student.Id == selectedStudent.Id)
+                .FirstOrDefault();
+            //this should never happen
+            if (pairedWatch == null)
+                UnknownFailureMessageBox();
+            else {
+                pairedWatch.Student = null;
+                RefreshAvailableWatchesListBox();
+                RefreshAvailableStudentsListBox();
+                RefreshPairedStudentsListBox();
+            }
+        }
     }
 }
